+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentFundamentalsSpec.scala
@@ -44,7 +44,8 @@ import tech.units.indriya.quantity.Quantities
 
 import java.util.UUID
 import java.util.concurrent.TimeUnit
-import javax.measure.quantity.{Energy, Power}
+import javax.measure.quantity.Power
+import scala.collection.SortedSet
 
 class ParticipantAgentFundamentalsSpec
     extends AgentSpec(
@@ -226,7 +227,7 @@ class ParticipantAgentFundamentalsSpec
                 operationEnd
               )
 
-            additionalActivationTicks.corresponds(expectedTicks.toArray)(
+            additionalActivationTicks.corresponds(expectedTicks)(
               _ == _
             ) shouldBe true
           }
@@ -237,7 +238,7 @@ class ParticipantAgentFundamentalsSpec
   "Determining the next activation tick" should {
     "bring up no activation trigger" in {
       val baseStateData = ParticipantAgentFundamentalsSpec.mockBaseStateData(
-        Array.emptyLongArray,
+        SortedSet.empty,
         Map.empty
       )
 
@@ -252,7 +253,7 @@ class ParticipantAgentFundamentalsSpec
 
     "bring up the next foreseen data tick, if this is the closest one" in {
       val baseStateData = ParticipantAgentFundamentalsSpec.mockBaseStateData(
-        Array(100L, 200L, 300L),
+        SortedSet(100L, 200L, 300L),
         Map(
           self -> Some(10L),
           noSender -> Some(0L)
@@ -283,7 +284,7 @@ class ParticipantAgentFundamentalsSpec
 
     "bring up the next additional activation tick, if this is the closest one and pop it from base state data" in {
       val baseStateData = ParticipantAgentFundamentalsSpec.mockBaseStateData(
-        Array(0L, 10L, 20L),
+        SortedSet(0L, 10L, 20L),
         Map(
           self -> Some(200L),
           noSender -> Some(100L)
@@ -317,7 +318,7 @@ class ParticipantAgentFundamentalsSpec
 
     "bring up the next additional activation tick, if it and a data tick at the same time are the closest ones and pop it from base state data" in {
       val baseStateData = ParticipantAgentFundamentalsSpec.mockBaseStateData(
-        Array(0L, 10L, 20L),
+        SortedSet(0L, 10L, 20L),
         Map(
           self -> Some(20L),
           noSender -> Some(0L)
@@ -602,7 +603,7 @@ class ParticipantAgentFundamentalsSpec
         ),
         None,
         outputConfig,
-        Array(0L, 900L, 1800L),
+        SortedSet(0L, 900L, 1800L),
         Map.empty,
         1e-12,
         ValueStore
@@ -635,7 +636,7 @@ class ParticipantAgentFundamentalsSpec
         ),
         None,
         outputConfig,
-        Array(0L, 900L, 1800L),
+        SortedSet(0L, 900L, 1800L),
         Map.empty,
         1e-12,
         ValueStore(901L),
@@ -664,7 +665,7 @@ case object ParticipantAgentFundamentalsSpec extends MockitoSugar {
     *   mocked base state data
     */
   def mockBaseStateData(
-      additionalActivationTicks: Array[Long],
+      additionalActivationTicks: SortedSet[Long],
       foreseenDataTicks: Map[ActorRef, Option[Long]]
   ): ParticipantModelBaseStateData[
     ApparentPower,